Reform code
