--- src/mux/client.rs
@@ -1,276 +1,0 @@
-//! Client side of the multiplexor
-//! SPDX-License-Identifier: Apache-2.0 OR GPL-3.0-or-later
-
-use super::frame::{DatagramFrame, Frame, StreamFlag, StreamFrame};
-use super::IntKey;
-use futures_util::{Sink as FutureSink, SinkExt, Stream as FutureStream, StreamExt};
-use std::collections::HashMap;
-use std::sync::Arc;
-pub use tokio::io::DuplexStream;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt, WriteHalf};
-use tokio::sync::RwLock;
-use tracing::{debug, error, trace, warn};
-use tungstenite::Message;
-
-/// All parameters of a stream channel
-#[derive(Debug)]
-pub struct MuxStream<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Unpin + 'static,
-{
-    /// Communication happens here
-    pub stream: DuplexStream,
-    /// Our port
-    pub port: u16,
-    /// Port of the other end
-    pub server_port: u16,
-    inner: Arc<MultiplexorInner<Sink, Stream>>,
-}
-
-impl<Sink, Stream> Drop for MuxStream<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Unpin + 'static,
-{
-    fn drop(&mut self) {
-        // Notify the task that this port is no longer in use
-        self.inner
-            .may_close_ports_tx
-            .send(self.port)
-            // Maybe the task has already exited, who knows
-            .unwrap_or_else(|_| warn!("Failed to notify task of dropped port"));
-    }
-}
-
-super::common_methods::make_asyncrw_proxy! {}
-
-/// Multiplexor inner
-#[derive(Debug)]
-pub struct MultiplexorInner<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Unpin + 'static,
-{
-    /// The underlying `Sink` of messages
-    sink: RwLock<Sink>,
-    /// The underlying `Stream` of messages
-    stream: RwLock<Stream>,
-    /// Interval between keepalive `Ping`s
-    keepalive_interval: Option<std::time::Duration>,
-    /// Open stream channels
-    streams: RwLock<HashMap<u16, WriteHalf<DuplexStream>>>,
-    /// Channel of received datagram frames for processing
-    datagram_rx: RwLock<tokio::sync::mpsc::Receiver<DatagramFrame>>,
-    /// Channel of established streams for processing
-    stream_rx: RwLock<tokio::sync::mpsc::Receiver<MuxStream<Sink, Stream>>>,
-    /// Channel for notifying the task of a dropped port
-    may_close_ports_tx: tokio::sync::mpsc::UnboundedSender<u16>,
-}
-
-#[derive(Clone, Debug)]
-pub struct Multiplexor<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Unpin + 'static,
-{
-    inner: Arc<MultiplexorInner<Sink, Stream>>,
-}
-
-impl<Sink, Stream> Multiplexor<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Sync + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Sync + Unpin + 'static,
-{
-    /// Create a new `Multiplexor`
-    #[tracing::instrument(skip_all, level = "debug")]
-    pub fn new(
-        sink: Sink,
-        stream: Stream,
-        keepalive_interval: Option<std::time::Duration>,
-    ) -> Self {
-        let (datagram_tx, datagram_rx) = tokio::sync::mpsc::channel(super::DATAGRAM_BUFFER_SIZE);
-        let (stream_tx, stream_rx) = tokio::sync::mpsc::channel(super::STREAM_BUFFER_SIZE);
-        let (may_close_ports_tx, may_close_ports_rx) = tokio::sync::mpsc::unbounded_channel();
-
-        let inner = Arc::new(MultiplexorInner {
-            sink: RwLock::new(sink),
-            stream: RwLock::new(stream),
-            keepalive_interval,
-            streams: RwLock::new(HashMap::new()),
-            datagram_rx: RwLock::new(datagram_rx),
-            stream_rx: RwLock::new(stream_rx),
-            may_close_ports_tx,
-        });
-
-        tokio::spawn(
-            inner
-                .clone()
-                .task(datagram_tx, stream_tx, may_close_ports_rx),
-        );
-        trace!("Multiplexor task spawned");
-
-        Self { inner }
-    }
-
-    /// Request a channel for `host` and `port`
-    /// Returns `None` if the connection is closed
-    #[tracing::instrument(skip(self), level = "debug")]
-    pub async fn new_stream_channel(
-        &self,
-        host: Vec<u8>,
-        port: u16,
-    ) -> Result<MuxStream<Sink, Stream>, super::Error> {
-        let host_len = host.len();
-        let mut syn_payload =
-            Vec::with_capacity(std::mem::size_of::<u8>() + std::mem::size_of::<u16>() + host_len);
-        syn_payload.push(u8::try_from(host_len)?);
-        syn_payload.extend_from_slice(&host);
-        syn_payload.extend_from_slice(&port.to_be_bytes());
-        // Allocate a new port
-        let sport = u16::next_available_key(&*self.inner.streams.read().await);
-        trace!("sport = {sport}");
-        let syn_frame = StreamFrame {
-            sport,
-            dport: 0,
-            flag: StreamFlag::Syn,
-            data: syn_payload,
-        };
-        trace!("sending syn");
-        self.inner.send_frame(Frame::Stream(syn_frame)).await?;
-        trace!("sending stream to user");
-        let mut stream_rx = self.inner.stream_rx.write().await;
-        // `unwrap`. Panic implies my logic is wrong
-        let stream = stream_rx.recv().await.unwrap();
-        Ok(stream)
-    }
-
-    /// Get the next available datagram
-    /// Returns `None` if the connection is closed
-    #[tracing::instrument(skip(self), level = "debug")]
-    pub async fn get_datagram(&self) -> Option<DatagramFrame> {
-        let mut datagram_rx = self.inner.datagram_rx.write().await;
-        datagram_rx.recv().await
-    }
-
-    /// Send a datagram
-    #[tracing::instrument(skip(self), level = "debug")]
-    pub async fn send_datagram(&self, frame: DatagramFrame) -> Result<(), super::Error> {
-        self.inner.send_frame(Frame::Datagram(frame)).await?;
-        Ok(())
-    }
-}
-
-impl<Sink, Stream> MultiplexorInner<Sink, Stream>
-where
-    Stream: FutureStream<Item = tungstenite::Result<Message>> + Send + Sync + Unpin + 'static,
-    Sink: FutureSink<Message, Error = tungstenite::Error> + Send + Sync + Unpin + 'static,
-{
-    /// Client-side processing task
-    #[tracing::instrument(skip_all, level = "trace")]
-    pub async fn task(
-        self: Arc<Self>,
-        mut datagram_tx: tokio::sync::mpsc::Sender<DatagramFrame>,
-        mut stream_tx: tokio::sync::mpsc::Sender<MuxStream<Sink, Stream>>,
-        mut may_close_ports_rx: tokio::sync::mpsc::UnboundedReceiver<u16>,
-    ) -> Result<(), super::Error> {
-        let keepalive_interval = self
-            .keepalive_interval
-            .unwrap_or(tokio::time::Duration::MAX);
-        // This is a bit of a hack, but there seems to be no way to optionally unwrap in `select!`
-        let mut keepalive_interval = tokio::time::interval(keepalive_interval);
-        loop {
-            tokio::select! {
-                Some(port) = may_close_ports_rx.recv() => {
-                    debug!("freeing port: {}", port);
-                    self.streams.write().await.remove(&port);
-                }
-                mut stream = self.stream.write() => {
-                    let msg = stream.next().await;
-                    if let Some(msg) = msg {
-                        let msg = msg?;
-                        trace!("received message: {:?}", msg);
-                        if self.clone().process_message(msg, &mut datagram_tx, &mut stream_tx).await? {
-                            break;
-                        }
-                    } else {
-                        break;
-                    }
-                }
-                _ = keepalive_interval.tick() => {
-                    let mut sink = self.sink.write().await;
-                    trace!("sending ping");
-                    sink.send(Message::Ping(vec![])).await?;
-                }
-                else => {
-                    break;
-                }
-            }
-        }
-        self.close_all_write().await;
-        Ok(())
-    }
-
-    /// Process a stream frame
-    /// Does the following:
-    /// - If `flag` is `Ack`, we create a `DuplexStream` and send it to the
-    ///  `stream_tx` channel.
-    /// - Otherwise, we find the `DuplexStream` with the matching `dport` and
-    ///   - Send the data to the `DuplexStream`.
-    ///   - If the `DuplexStream` is closed, send a `Fin` frame.
-    #[tracing::instrument(skip_all, level = "trace")]
-    async fn process_stream_frame(
-        self: Arc<Self>,
-        stream_frame: StreamFrame,
-        stream_tx: &mut tokio::sync::mpsc::Sender<MuxStream<Sink, Stream>>,
-    ) -> Result<(), super::Error> {
-        let port = stream_frame.dport;
-        match stream_frame.flag {
-            StreamFlag::Syn => {
-                panic!("`Syn` flag should not be received by client");
-            }
-            StreamFlag::Ack => {
-                // See `server.rs`
-                let (our, their) = tokio::io::duplex(super::DUPLEX_SIZE);
-                let (our_rx, our_tx) = tokio::io::split(our);
-                // Save the TX end of the stream so we can write to it when subsequent frames arrive
-                let mut streams = self.streams.write().await;
-                streams.insert(port, our_tx);
-                drop(streams);
-                let stream = MuxStream {
-                    stream: their,
-                    port,
-                    server_port: stream_frame.sport,
-                    inner: self.clone(),
-                };
-                // This goes to the user
-                stream_tx
-                    .send(stream)
-                    .await
-                    .map_err(|e| super::Error::SendStreamToClient(e.to_string()))?;
-                // Read from the stream and pack frames
-                tokio::spawn(self.stream_task(port, stream_frame.sport, our_rx));
-            }
-            StreamFlag::Fin => {
-                self.close_write(port).await;
-            }
-            StreamFlag::Psh => {
-                let mut streams = self.streams.write().await;
-                if let Some(stream) = streams.get_mut(&port) {
-                    stream.write_all(&stream_frame.data).await?;
-                }
-            }
-        }
-        Ok(())
-    }
-
-    super::common_methods::make_send_frame! {}
-
-    super::common_methods::make_process_message! {}
-
-    super::common_methods::make_close_all_write! {}
-
-    super::common_methods::make_close_write! {}
-
-    super::common_methods::make_stream_task! {}
-}
